+++ packages/address-picker/src/AddressPicker.example.tsx
@@ -1,15 +1,24 @@
 import * as React from 'react';
 import AddressPicker from './index';
 
-export default () => (
-  <AddressPicker
-    onChange={(e) => {
-      // console.log(e);
-    }}
-    onSelect={(e) => {
-      // console.log(e);
-    }}
-    showMap
-    country="au"
-  />
-);
+export default class AddressPickerExample extends React.Component {
+  state = {
+    address: '',
+  };
+
+  render() {
+    return (
+      <>
+        <AddressPicker
+          onChange={() => undefined}
+          onSelect={(e) => {
+            this.setState(e);
+          }}
+          showMap
+          country="au"
+        />
+        <pre>{JSON.stringify(this.state, null, 2)}</pre>
+      </>
+    );
+  }
+}