+++ packages/address-picker/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog - @lendi-ui/address-picker
 
+## [1.1.0] 2019-6-27 [HUB-66](https://creditandfinance.atlassian.net/browse/HUB-66)
+### Created
+- Added country to the addressObject
+- Fixed bug with onSelect not returning the correct response
+- Added isDisabled and size to the props
+
 ## [1.0.0] 2019-6-27 [HUB-66](https://creditandfinance.atlassian.net/browse/HUB-66)
 ### Created
 - Release version of AddressPicker.